Alteração do tipo dos atributos relativos à temperatura de inteiro para ponto flutuante.

+++ src/main/java/net/eldiosantos/gobocom/arcondicionado/controller/Controller.java
@@ -28,6 +28,6 @@ public interface Controller {
 	 *            Tempo em minutos.
 	 * @return Retorna o custo da chamada.
 	 */
-	public abstract BigDecimal manterTemperatura(int tempInicial,
-			int tempFinal, int minutos);
+	public abstract BigDecimal manterTemperatura(float tempInicial,
+			float tempFinal, int minutos);
 }
